Round_1B/requirements.txt
fitz
PyMuPDF==1.23.19
scikit-learn
pandas
numpy
faiss-cpu
sentence-transformers
